+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- It is now possible to use ajv instance to extend schema. Use fastify addSchema feature is not supported because the validator compiler is overwritten.
+- It is now possible to use ajv instance to extend schema using `addValidatorSchema`. Use fastify addSchema feature is not supported because the validator compiler is overwritten.
+- Add new function `getValidatorSchema` to get schema added to the validator.
 
 ## v4.0.0 - 2021-08-31